Code cleanup
Cleaned up imports, comments, and small things that should be fixed
